+++ include/Events/BalanceChangeEventData.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+#include "EventDataInterface.hpp"
+
+class BalanceChangeEventData : public EventDataInterface{
+public:
+  int balanceChange;
+  float timeStamp;
+
+  BalanceChangeEventData(int balanceChange, float timeStamp);
+};