+++ src/Events/LevelChangeEventData.hpp
@@ -1,0 +1,6 @@
+#include "Events/LevelChangeEventData.hpp"
+
+LevelChangeEventData::LevelChangeEventData(int level, float timeStamp){
+  this -> timeStamp = timeStamp;
+  this -> level = level;
+}