+++ src/Events/OptionSelectedEventData.cpp
@@ -1,7 +1,7 @@
 #include "Events/OptionSelectedEventData.hpp"
 
-OptionSelectedEventData::OptionSelectedEventData(int optionID, float timeStamp){
+OptionSelectedEventData::OptionSelectedEventData(int optionID, int newValue, float timeStamp){
   this -> timeStamp = timeStamp;
   this -> optionID = optionID;
-
+  this -> newValue = newValue;
 }