+++ src/UserView.cpp
@@ -78,7 +78,7 @@ void UserView::initScreens(){
   //Buy Tower Sreeen
   shared_ptr<Screen> buyTowerScreen = make_shared<BuyTowerScreen>(eventManager, textLoader,gameLogic, windowX, windowY);
   //Playing Screen
-  shared_ptr<Screen> playingScreen = make_shared<PlayingScreen>(eventManager, textLoader,gameLogic, buyTowerScreen, windowX, windowY);
+  shared_ptr<Screen> playingScreen = make_shared<PlayingScreen>(eventManager, textLoader,gameLogic, windowX, windowY);
   //Restart Screen
   shared_ptr<Screen> restartScreen = make_shared<RestartScreen>(eventManager, textLoader,windowX, windowY);
 
@@ -190,6 +190,21 @@ void UserView::handleStateChange(const EventInterface& event){
 
   //change the screen to the current state
   screen = state;
+
+  //we have to check the row and col if this is the buyTowerScreen
+  if(state == State::BuyTower){
+    int row = scEventData -> row;
+    int col = scEventData -> col;
+    assert(row < gameLogic->getRows());
+    assert(row >= 0);
+    assert(col < gameLogic->getCols());
+    assert(col >=0);
+
+    //cast the screen selected(which should be buy tower)
+    BuyTowerScreen* buyTowerScreen = dynamic_cast<BuyTowerScreen*>(((this->screens).at((int)screen)).get());
+
+    buyTowerScreen -> setSelectedTile(row,col);
+  }
 
   //register all the delegagtes for the new state
   (this->screens).at((int)screen) -> registerDelegates();